+++ webui/src/i18n/index.ts
@@ -1,9 +1,0 @@
-import { en } from "./en";
-import { zh } from "./zh";
-
-export type TranslationKeys = keyof typeof en;
-
-export {
-  en,
-  zh,
-};