+++ webui/tailwind.config.d.ts
@@ -1,8 +1,0 @@
-declare const _default: {
-    content: string[];
-    theme: {
-        extend: {};
-    };
-    plugins: any[];
-};
-export default _default;